requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
matplotlib==3.9.0
fpdf2==2.7.9
reportlab==4.2.0
openpyxl==3.1.4
python-dotenv==1.0.1
gunicorn==22.0.0
